+++ astrality/tests/event_listener/test_daylight.py
@@ -11,8 +11,8 @@ def daylight_config():
     """A daylight event listener configuration in Trondheim, Norway."""
     return {
         'type': 'daylight',
-        'latitude': 63.446827,
-        'longitude': 10.421906,
+        'latitude': 0,
+        'longitude': 0,
         'elevation': 0,
     }